Refactor Breadcrumb components to allow for usage with route configuration.
